malloc using word alignment
Added a new example to demonstrate how to use malloc and free if we want
our array to use 2 bytes instead of 4 for the alignment.

We still should save the eax register, but we save it in the first 4
bytes (occupying 2 addresses). Starting from address mem+4, we can use
our array to store 2 bytes of memory.

The same concept should apply if we want to use a memory with 1 byte
pagination, but in that case, the eax register will have to be the first
4 elements of the array.

If we want to use a quad word or 8 bytes, we can simply store the whole
rax in the first address.